Unify push notification router
